Modified:ElementCollection And Excepation Handling

+++ src/main/java/com/bridgelabz/addressbookapp/service/AddressService.java
@@ -2,6 +2,7 @@ package com.bridgelabz.addressbookapp.service;
 
 import com.bridgelabz.addressbookapp.dto.AddressBookDTO;
 import com.bridgelabz.addressbookapp.entity.Address;
+import com.bridgelabz.addressbookapp.exception.AddressBookException;
 import com.bridgelabz.addressbookapp.repository.AddressRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -19,32 +20,37 @@ public class AddressService implements IAddressBookService {
         repository.save(newAddress);
         return newAddress;
     }
-
-
-    public List<Address> getRecordByFirstName(String firstName) {
-        return repository.findByFirstName(firstName);
-    }
-    public Optional<Address> getRecordById(Integer id) {
-        Optional<Address> newBook = repository.findById(id);
-        return newBook;
-    }
-
-    public List<Address> getRecordByName() {
-        return repository.findAll();
+    //Created service method which serves controller api to get record by id
+    public Address getRecordById(Integer id) {
+        Optional<Address> address = repository.findById(id);
+        if (address.isPresent()) {
+            return address.get();
+        } else throw new AddressBookException("Addressbook id not found");
     }
 
     public List<Address> getRecord() {
         return repository.findAll();
     }
 
+    //Created service method which serves controller api to update record by id
     public Address updateRecordById(Integer id, AddressBookDTO addressBookDTO) {
+        Optional<Address> addressBook = repository.findById(id);
+        if (addressBook.isEmpty()) {
+            throw new AddressBookException("AddressBook details for id not found");
+        }
         Address newBook = new Address(id, addressBookDTO);
         repository.save(newBook);
         return newBook;
     }
 
+    //Created service method which serves controller api to delete record by id
     public String deleteRecordById(Integer id) {
-        repository.deleteById(id);
+        Optional<Address> newAddressBook = repository.findById(id);
+        if (newAddressBook.isEmpty()) {
+            throw new AddressBookException("Address Book Details not found");
+        } else {
+            repository.deleteById(id);
+        }
         return null;
     }
